Refactor its tests.
